refactor: move unit tests to tests/Unit and update configuration
