+++ phpunit.xml
@@ -6,19 +6,12 @@
          cacheDirectory=".phpunit.cache">
     <testsuites>
         <testsuite name="unit">
-            <directory>tests</directory>
-            <exclude>tests/Functional</exclude>
+            <directory>tests/Unit</directory>
         </testsuite>
         <testsuite name="functional">
             <directory>tests/Functional</directory>
         </testsuite>
     </testsuites>
-
-    <groups>
-        <exclude>
-            <group>functional</group>
-        </exclude>
-    </groups>
 
     <source>
         <include>
