+++ tests/Unit/Config/YamlConfigProviderTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tests\Config;
+namespace Tests\Unit\Config;
 
 use Igancev\WorkReporter\Config\ConfigException;
 use Igancev\WorkReporter\Config\YamlConfigProvider;
