+++ tests/Unit/InitCommandTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tests;
+namespace Tests\Unit;
 
 use Igancev\WorkReporter\InitCommand;
 use PHPUnit\Framework\Attributes\CoversClass;
