+++ tests/Unit/Source/SuperProductivity/StorageTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tests\Source\SuperProductivity;
+namespace Tests\Unit\Source\SuperProductivity;
 
 use Igancev\WorkReporter\Source\SuperProductivity\Storage;
 use Igancev\WorkReporter\Source\SuperProductivity\Tag;
