+++ tests/Unit/TimeEntryCollectionTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tests;
+namespace Tests\Unit;
 
 use DateTimeImmutable;
 use Igancev\WorkReporter\Duration;
