+++ tests/Unit/VersionTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tests;
+namespace Tests\Unit;
 
 use Igancev\WorkReporter\Version;
 use PHPUnit\Framework\Attributes\CoversClass;
